This is the shittiest fucking thing ever. 8 hours down the drain and still can't add candidates to parties despite using the same configuration as a project where it does work

+++ src/main/java/localelection/programming24hr/repositories/CandidateRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Repository
 public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
-    Candidate findCandidateById(int id);
-    List<Candidate> deleteCandidateById(int id);
+    Candidate findCandidateById(Long id);
+    List<Candidate> deleteCandidateById(Long id);
     List<Candidate> findCandidateByFirstName(String firstName);
 }
